backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Celery and Messaging
celery==5.3.6
redis==5.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.3
tabulate==0.10.0

# AI/ML
google-generativeai
langchain-google-genai
langchain-groq==1.1.2
langchain-experimental==0.4.1
langchain-community==0.4.1
anthropic==0.18.1

# Docker
docker==7.0.0

# Validation
pydantic==2.12.5
pydantic-settings==2.13.1
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.28.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
